refactor: flatten reactive/ modules into the top-level pyjinhx package
Completes the consolidation: core/, reactive/, and config/ are now one flat
package of 16 cohesive modules (down from 34). Public API and behavior
unchanged; integrations/, builtins/, runtime/ retained as subpackages.
