+++ tests/reactive_test_support.py
@@ -7,8 +7,8 @@ from collections.abc import Iterator
 from contextlib import contextmanager
 
 from pyjinhx import FastAPIClientBackend
-from pyjinhx.reactive.client import ClientBackend
-from pyjinhx.reactive.mutations import MutationTracker
+from pyjinhx.client import ClientBackend
+from pyjinhx.mutations import MutationTracker
 
 
 class FakeRequest:
